Add ESP Accelerator Socket and Wrapper
This adds a Chisel ESP Accelerator socket and a wrapper that converts
ports to what the ESP Accelerator socket wants.

This includes tests that the wrapper generates the expected port names by
comparing it with Columbia Verilog. Additionally, this tests a trivial
accelerator that reports it's done a parameterized number of cycles after
kicked.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>